+++ apps/minifront/src/fetchers/balances/index.ts
@@ -5,7 +5,7 @@ import {
 import { AssetId } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
 import { AddressIndex } from '@penumbra-zone/protobuf/penumbra/core/keys/v1/keys_pb';
 import { ViewService } from '@penumbra-zone/protobuf';
-import { penumbra } from '../../prax';
+import { penumbra } from '../../penumbra';
 
 export interface BalancesProps {
   accountFilter?: AddressIndex;