+++ apps/minifront/src/fetchers/chain-id.ts
@@ -1,5 +1,5 @@
 import { ViewService } from '@penumbra-zone/protobuf';
-import { penumbra } from '../prax';
+import { penumbra } from '../penumbra';
 
 export const getChainId = async (): Promise<string | undefined> => {
   const { parameters } = await penumbra.service(ViewService).appParameters({});