+++ apps/minifront/vite.config.ts
@@ -2,7 +2,6 @@
 
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react-swc';
-import basicSsl from '@vitejs/plugin-basic-ssl';
 import { commitInfoPlugin } from './src/utils/commit-info-vite-plugin';
 import polyfillNode from 'vite-plugin-node-stdlib-browser';
 import svgr from 'vite-plugin-svgr';
@@ -15,7 +14,6 @@ export default defineConfig(({ mode }) => {
     plugins: [
       polyfillNode(),
       react(),
-      basicSsl(),
       commitInfoPlugin(),
       svgr({
         include: '**/*.svg',