+++ packages/getters/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @penumbra-zone/getters
 
+## 20.0.0
+
+### Patch Changes
+
+- Updated dependencies [3269282]
+  - @penumbra-zone/protobuf@6.3.0
+  - @penumbra-zone/bech32m@10.0.0
+
+## 19.0.0
+
+### Patch Changes
+
+- Updated dependencies [e543db4]
+  - @penumbra-zone/protobuf@6.2.0
+  - @penumbra-zone/bech32m@9.0.0
+
 ## 18.0.0
 
 ### Patch Changes