+++ packages/query/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @penumbra-zone/query
 
+## 36.0.0
+
+### Patch Changes
+
+- Updated dependencies [3269282]
+  - @penumbra-zone/protobuf@6.3.0
+  - @penumbra-zone/bech32m@10.0.0
+  - @penumbra-zone/getters@20.0.0
+  - @penumbra-zone/types@26.0.0
+  - @penumbra-zone/wasm@31.0.0
+  - @penumbra-zone/crypto-web@27.0.0
+
+## 35.1.0
+
+### Minor Changes
+
+- 8c036f5: batching storage operations with promises
+
+## 35.0.0
+
+### Patch Changes
+
+- Updated dependencies [48725e3]
+  - @penumbra-zone/wasm@30.1.0
+
+## 34.0.0
+
+### Patch Changes
+
+- 8bf66ea: Remove validator info throttle
+- Updated dependencies [e0db143]
+- Updated dependencies [e543db4]
+  - @penumbra-zone/wasm@30.0.0
+  - @penumbra-zone/protobuf@6.2.0
+  - @penumbra-zone/bech32m@9.0.0
+  - @penumbra-zone/getters@19.0.0
+  - @penumbra-zone/types@25.0.0
+  - @penumbra-zone/crypto-web@26.0.0
+
+## 33.0.0
+
+### Minor Changes
+
+- 487e26d: perform idb genesis in the block processor
+
+### Patch Changes
+
+- Updated dependencies [735e22b]
+  - @penumbra-zone/wasm@29.1.0
+
 ## 32.0.0
 
 ### Minor Changes