+++ packages/types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @penumbra-zone/types
 
+## 26.0.0
+
+### Patch Changes
+
+- Updated dependencies [3269282]
+  - @penumbra-zone/protobuf@6.3.0
+  - @penumbra-zone/bech32m@10.0.0
+  - @penumbra-zone/getters@20.0.0
+
+## 25.0.0
+
+### Patch Changes
+
+- Updated dependencies [e543db4]
+  - @penumbra-zone/protobuf@6.2.0
+  - @penumbra-zone/bech32m@9.0.0
+  - @penumbra-zone/getters@19.0.0
+
 ## 24.0.0
 
 ### Minor Changes